requirements.txt
Django==5.0.3
httpx==0.26.0
python-dotenv==1.0.1
django-cors-headers==4.3.1 